Initial Creation of mkv2mp4
When invoked, mkv2mp4 scans the contents of the current working directory looking for .mkv and .srt
files. It attempts to match SRTs to the right MKV and then, for each video, creates an
MP4 (including the matched SRT, if possible).

If the input codecs are already acceptable (AAC for the audio, H264 for the video) they are copied
without modification.